Fix non robust test (issue1091-bogus-rebuild)
The test was susceptible to interference from other tests, because it
assumed the 1-exec-simple project was clean, which may be not. The
reason that made the test work until now was that it happened to execute
before other tests that use the same project.

Now it ensures the project is clean before running.

This commit is part of PR #1490 because it was needed in order to make
the new colored-output test pass (which was interfering with this one).